split workers.ts
