+++ cli/js/workers.ts
@@ -12,12 +12,12 @@ import { EventTarget } from "./event_target.ts";
 const encoder = new TextEncoder();
 const decoder = new TextDecoder();
 
-export function encodeMessage(data: any): Uint8Array {
+function encodeMessage(data: any): Uint8Array {
   const dataJson = JSON.stringify(data);
   return encoder.encode(dataJson);
 }
 
-export function decodeMessage(dataIntArray: Uint8Array): any {
+function decodeMessage(dataIntArray: Uint8Array): any {
   const dataJson = decoder.decode(dataIntArray);
   return JSON.parse(dataJson);
 }
@@ -65,80 +65,11 @@ async function hostGetMessage(id: number): Promise<any> {
   }
 }
 
-// Stuff for workers
-export const onmessage: (e: { data: any }) => void = (): void => {};
-export const onerror: (e: { data: any }) => void = (): void => {};
-
-export function postMessage(data: any): void {
-  const dataIntArray = encodeMessage(data);
-  sendSync(dispatch.OP_WORKER_POST_MESSAGE, {}, dataIntArray);
-}
-
-export async function getMessage(): Promise<any> {
-  log("getMessage");
-  const res = await sendAsync(dispatch.OP_WORKER_GET_MESSAGE);
-  if (res.data != null) {
-    return decodeMessage(new Uint8Array(res.data));
-  } else {
-    return null;
-  }
-}
-
-export let isClosing = false;
-
-export function workerClose(): void {
-  isClosing = true;
-}
-
-export async function workerMain(): Promise<void> {
-  log("workerMain");
-
-  while (!isClosing) {
-    const data = await getMessage();
-    if (data == null) {
-      log("workerMain got null message. quitting.");
-      break;
-    }
-
-    let result: void | Promise<void>;
-    const event = { data };
-
-    try {
-      if (!globalThis["onmessage"]) {
-        break;
-      }
-      result = globalThis.onmessage!(event);
-      if (result && "then" in result) {
-        await result;
-      }
-      if (!globalThis["onmessage"]) {
-        break;
-      }
-    } catch (e) {
-      if (globalThis["onerror"]) {
-        const result = globalThis.onerror(
-          e.message,
-          e.fileName,
-          e.lineNumber,
-          e.columnNumber,
-          e
-        );
-        if (result === true) {
-          continue;
-        }
-      }
-      throw e;
-    }
-  }
-}
-
 export interface Worker {
   onerror?: (e: any) => void;
   onmessage?: (e: { data: any }) => void;
   onmessageerror?: () => void;
   postMessage(data: any): void;
-  // TODO(bartlomieju): remove this
-  closed: Promise<void>;
 }
 
 export interface WorkerOptions {
@@ -150,7 +81,6 @@ export class WorkerImpl extends EventTarget implements Worker {
   private isClosing = false;
   private messageBuffer: any[] = [];
   private ready = false;
-  private readonly isClosedPromise: Resolvable<void>;
   public onerror?: (e: any) => void;
   public onmessage?: (data: any) => void;
   public onmessageerror?: () => void;
@@ -166,13 +96,14 @@ export class WorkerImpl extends EventTarget implements Worker {
 
     if (type !== "module") {
       throw new Error(
-        'Not yet implemented: only type: "module" workers are supported'
+        'Not yet implemented: only "module" type workers are supported'
       );
     }
 
     let hasSourceCode = false;
     let sourceCode = new Uint8Array();
 
+    /* TODO(bartlomieju):
     // Handle blob URL.
     if (specifier.startsWith("blob:")) {
       hasSourceCode = true;
@@ -186,16 +117,12 @@ export class WorkerImpl extends EventTarget implements Worker {
       }
       sourceCode = blobBytes!;
     }
+    */
 
     const { id, loaded } = createWorker(specifier, hasSourceCode, sourceCode);
     this.id = id;
     this.ready = loaded;
-    this.isClosedPromise = createResolvable();
     this.poll();
-  }
-
-  get closed(): Promise<void> {
-    return this.isClosedPromise;
   }
 
   private handleError(e: any): boolean {
@@ -253,7 +180,6 @@ export class WorkerImpl extends EventTarget implements Worker {
       } else {
         this.isClosing = true;
         hostCloseWorker(this.id);
-        this.isClosedPromise.resolve();
         break;
       }
     }
